% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abart_fix.R
\name{abart.fix}
\alias{abart.fix}
\title{fixes bug in abart program from the BART package}
\usage{
abart.fix(
  x.train,
  times,
  delta,
  x.test = matrix(0, 0, 0),
  K = 100,
  type = "abart",
  ntype = 1,
  sparse = FALSE,
  theta = 0,
  omega = 1,
  a = 0.5,
  b = 1,
  augment = FALSE,
  rho = NULL,
  xinfo = matrix(0, 0, 0),
  usequants = FALSE,
  rm.const = TRUE,
  sigest = NA,
  sigdf = 3,
  sigquant = 0.9,
  k = 2,
  power = 2,
  base = 0.95,
  lambda = NA,
  tau.num = c(NA, 3, 6)[ntype],
  offset = NULL,
  w = rep(1, length(times)),
  ntree = c(200L, 50L, 50L)[ntype],
  numcut = 100L,
  ndpost = 1000L,
  nskip = 100L,
  keepevery = c(1L, 10L, 10L)[ntype],
  printevery = 100L,
  transposed = FALSE,
  mc.cores = 1L,
  nice = 19L,
  seed = 99L
)
}
\description{
fixes bug in abart program from the BART package
}
\details{
For arguments and such see \link[BART]{abart}.
}
