% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/construct_matrix.R
\name{construct_interp_survival_matrix}
\alias{construct_interp_survival_matrix}
\title{Constructs survival matrix for use in applied data examples}
\usage{
construct_interp_survival_matrix(x, nT, intercept = FALSE)
}
\arguments{
\item{x}{covariates}

\item{nT}{number of unique times}

\item{intercept}{include an intercept?}
}
\value{

}
\description{
Constructs survival matrix for use in applied data examples
}
