% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/experimentWPmethod.R
\name{experimentWPMethod}
\alias{experimentWPMethod}
\title{Simulation study function}
\usage{
experimentWPMethod(target, hyperparameters, conditions)
}
\arguments{
\item{target}{function that is the output of the get_binary_nonlinear, get_normal_linear, etc}

\item{hyperparameters}{hyperparameters for the Bayesian methodology}

\item{conditions}{conditions for the simulation}
}
\value{

}
\description{
Simulation study function
}
