% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gam_iterate_gbm.R
\name{gam_iterate}
\alias{gam_iterate}
\title{Run the p-Wasserstein projections for the GAM example}
\usage{
gam_iterate(
  formula,
  y,
  x,
  extract,
  time,
  nT,
  which.gam = c("gam", "qgam"),
  ...
)
}
\arguments{
\item{formula}{an R formula object suitable for \link[gam]{gam} or \link[qgam]{qgam}}

\item{y}{prediction matrix}

\item{x}{covariates}

\item{extract}{new data for prediction}

\item{time}{vector of times}

\item{nT}{number of timepoints}

\item{which.gam}{"gam" or "qgam" for the models?}

\item{...}{not currently used}
}
\value{

}
\description{
Run the p-Wasserstein projections for the GAM example
}
