% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gam_iterate_gbm.R
\name{gamm_interp_data_gbm}
\alias{gamm_interp_data_gbm}
\title{Title}
\usage{
gamm_interp_data_gbm(gammX, times)
}
\arguments{
\item{gammX}{gam model}

\item{times}{event times}
}
\value{

}
\description{
Title
}
