% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_binary_nonlinear.R
\name{get_binary_nonlinear_model}
\alias{get_binary_nonlinear_model}
\title{Get the binary nonlinear data simulation}
\usage{
get_binary_nonlinear_model()
}
\value{
list with slots
`rprior`
`rdata`
`rpost`
`X = list(rX = rX, corr = NULL)`, rX being a function to simulate the x data = function(n, corr, p)
`data_gen_functions`
 `rparam`
 `sel.pred.fun `
 `link` with the logistic link function
 `invlink` with the logistic inverse link function
 `mf.data`
}
\description{
Get the binary nonlinear data simulation
}
