% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_normal_linear.R
\name{get_normal_linear_model}
\alias{get_normal_linear_model}
\title{Get the Gaussian linear data simulation}
\usage{
get_normal_linear_model()
}
\value{
list with slots
`rprior`
`rdata`
`rpost`
`X = list(rX = rX, corr = NULL)`, rX being a function to simulate the x data = function(n, corr, p)
 `data_gen_functions`
 `rparam`
 `sel.pred.fun `
 `link` with the logistic link function
 `invlink` with the logistic inverse link function
}
\description{
Get the Gaussian linear data simulation
}
