% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/log_sum_exp.R
\name{log_sum_exp}
\alias{log_sum_exp}
\title{safely sums values on the exp scale from the log scale}
\usage{
log_sum_exp(x)
}
\arguments{
\item{x}{vector to be summed}
}
\value{
real number
}
\description{
safely sums values on the exp scale from the log scale
}
