% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/log_sum_exp.R
\name{log_sum_exp2}
\alias{log_sum_exp2}
\title{logged sum of exponentiated variables}
\usage{
log_sum_exp2(x, y)
}
\arguments{
\item{x}{a number}

\item{y}{another number}
}
\value{
\eqn{\log(\exp(x) + \exp(y))}
}
\description{
logged sum of exponentiated variables
}
