% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.compare.R
\name{plot.compare}
\alias{plot.compare}
\title{Plot to compare WpProj outputs}
\usage{
\method{plot}{compare}(
  models,
  target = NULL,
  X = NULL,
  theta = NULL,
  method = c("w2", "mse"),
  quantity = c("posterior", "mean"),
  parallel = FALSE,
  transform = function(x) {     return(x) }
)
}
\arguments{
\item{models}{WpProj models}

\item{target}{target predicctions}

\item{X}{covariates}

\item{theta}{coefficients}

\item{method}{w2 or mse?}

\item{quantity}{posterior (parameters) or mean}

\item{parallel}{foreach backend}

\item{transform}{transform to use}
}
\value{

}
\description{
Plot to compare WpProj outputs
}
