% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survival_to_poisson.R
\name{pois_dat}
\alias{pois_dat}
\title{Constructs Poisson data for use in Survival models}
\usage{
pois_dat(time, event.times, event, x)
}
\arguments{
\item{time}{observation times}

\item{event.times}{times at which events coccur}

\item{event}{indicator for censored (0) or event (1)}

\item{x}{covariates}
}
\value{

}
\description{
Constructs Poisson data for use in Survival models
}
